`addStrings` method to keep track of extracted strings
This introduces state to the parser to store strings that have
been extracted. This is necessary for tools that process multiple
files and want all strings stored in one catalog with references,
comments, etc. merged.

This commit also adds a method to convert the saved strings into
an array of PO files.

Closes #6, #9