Allow extracting non-literals
This commit allows extraction of the form `{{_ foo}}` where `foo`
is *not* a string literal. This is so we don't blow up in the
face of apps using gettext incorrectly and not having a real
msgid. We do log a warning in this case though, to prod you in
the right direction.